refactor: extract lyrics editor helpers to separate module
